+++ src/store/mod.rs
@@ -1,10 +1,13 @@
 pub mod memory;
 
+mod store_trait;
+pub use store_trait::*;
+
 #[cfg(feature = "postgres-store")]
 pub mod postgres;
 
 #[cfg(feature = "redis-store")]
 pub mod redis;
 
-mod store_trait;
-pub use store_trait::*;
+#[cfg(feature = "layered-store")]
+pub mod layered;